requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
deep-sort-realtime==1.3.2
Deprecated==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
einops==0.8.0
filelock==3.17.0
filterpy==1.4.5
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
gyp==0.1
httplib2==0.20.2
huggingface-hub==0.27.1
hyperlink==21.0.0
idna==3.3
imageio==2.37.0
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libvirt-python==8.0.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.10.1
matrix-client==0.4.0
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycurl==7.44.1
PyGithub==2.5.0
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.25.1
rich==13.9.4
safetensors==0.5.2
scikit-image==0.25.2
scipy==1.15.1
seaborn==0.13.2
SecretStorage==3.3.1
sentencepiece==0.2.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
soundfile==0.13.1
spandrel==0.4.1
ssh-import-id==5.11
sympy==1.13.1
systemd-python==234
tifffile==2025.3.30
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchsde==0.2.6
torchvision==0.20.1+cu121
tqdm==4.67.1
trampoline==0.1.2
transformers==4.48.1
triton==3.1.0
Twisted==22.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.107
ultralytics-thop==2.0.14
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websocket==0.2.1
wrapt==1.17.2
yarl==1.18.3
zipp==1.0.0
zope.event==5.0
zope.interface==5.4.0
